requirements.txt
pip~=23.0.1
django~=5.2.9
asgiref~=3.11.0
sqlparse~=0.5.5
setuptools~=65.5.0
djangorestframework==3.16.1
djangorestframework-simplejwt
djangorestframework-api-key